factor out mock s3
